fix: Moved around methods
